backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
bcrypt
email-validator
python-dotenv

